requirements.txt
requests
streamlit
pandas
numpy
datetime
python-dotenv
flask
fastapi
uvicorn
python-multipart
